Home Page kafka entities table UI fixes UI, refactored
