+++ ui-ngx/src/app/modules/home/components/kafka-tables/kafka-brokers-table.component.html
@@ -15,8 +15,8 @@
     limitations under the License.
 
 -->
-<div fxFlex class="tb-home-card-container tb-home-padding">
-  <tb-card-title-button style="padding: 10px 0 10px" [cardType]="cardType" [disabled]="true"></tb-card-title-button>
+<div fxFlex fxLayout="column" fxLayoutGap="12px" class="tb-home-card-container tb-home-padding">
+  <tb-card-title-button [cardType]="cardType" [disabled]="true"></tb-card-title-button>
   <section class="tb-kafka-table">
     <table mat-table [dataSource]="dataSource">
       <ng-container *ngFor="let column of columns" [matColumnDef]="column.key">
