+++ ui-ngx/src/app/modules/home/components/kafka-tables/kafka-tables-tab-group.component.ts
@@ -16,6 +16,9 @@
 
 import { ChangeDetectorRef, Component, OnInit } from '@angular/core';
 import { KafkaTable, KafkaTableTranslationMap } from '@shared/models/kafka.model';
+import { KafkaTopicsTableConfig } from '@home/components/kafka-tables/kafka-topics-table-config';
+import { KafkaService } from '@core/http/kafka.service';
+import { KafkaConsumerGroupsTableConfig } from '@home/components/kafka-tables/kafka-consumer-groups-table-config';
 
 @Component({
   selector: 'tb-kafka-tables-tab-group',
@@ -24,15 +27,20 @@ import { KafkaTable, KafkaTableTranslationMap } from '@shared/models/kafka.model
 })
 export class KafkaTablesTabGroupComponent implements OnInit {
 
+  public kafkaTopicsTableConfig: KafkaTopicsTableConfig;
+  public kafkaConsumerGroupsTableConfig: KafkaConsumerGroupsTableConfig;
   public selectedTab: KafkaTable;
   public readonly KafkaTable = KafkaTable;
   public readonly kafkaTableTranslationMap = KafkaTableTranslationMap;
 
-  constructor(private cd: ChangeDetectorRef) {
+  constructor(private kafkaService: KafkaService,
+              private cd: ChangeDetectorRef) {
   }
 
   ngOnInit() {
     this.selectedTab = KafkaTable.TOPICS;
+    this.kafkaTopicsTableConfig = new KafkaTopicsTableConfig(this.kafkaService);
+    this.kafkaConsumerGroupsTableConfig = new KafkaConsumerGroupsTableConfig(this.kafkaService);
   }
 
   select(table: KafkaTable) {
