+++ ui-ngx/src/app/modules/home/components/kafka-tables/kafka-topics-table-config.ts
@@ -15,17 +15,16 @@
 ///
 
 import { EntityTableColumn, EntityTableConfig } from '@home/models/entity/entities-table-config.models';
-import { TranslateService } from '@ngx-translate/core';
 import { TimePageLink } from '@shared/models/page/page-link';
 import { Observable } from 'rxjs';
 import { PageData } from '@shared/models/page/page-data';
-import { KafkaConsumerGroup } from '@shared/models/kafka.model';
+import { KafkaTopic } from '@shared/models/kafka.model';
 import { KafkaService } from '@core/http/kafka.service';
+import { formatBytes } from '@home/components/entity/entities-table-home.component';
 
-export class KafkaConsumerGroupsTableConfig extends EntityTableConfig<KafkaConsumerGroup, TimePageLink> {
+export class KafkaTopicsTableConfig extends EntityTableConfig<KafkaTopic, TimePageLink> {
 
   constructor(private kafkaService: KafkaService,
-              private translate: TranslateService,
               public entityId: string = null) {
     super();
     this.detailsPanelEnabled = false;
@@ -33,21 +32,26 @@ export class KafkaConsumerGroupsTableConfig extends EntityTableConfig<KafkaConsu
     this.addEnabled = false;
     this.entitiesDeleteEnabled = false;
     this.entityTranslations = {
-      noEntities: 'kafka.no-kafka-consumer-group-text',
-      search: 'kafka.consumer-groups-search'
+      noEntities: 'kafka.no-kafka-topic-text',
+      search: 'kafka.topics-search'
     };
 
-    this.entitiesFetchFunction = pageLink => this.fetchSessions(pageLink);
+    this.entitiesFetchFunction = pageLink => this.fetchKafkaTopics(pageLink);
 
     this.columns.push(
-      new EntityTableColumn<KafkaConsumerGroup>('groupId', 'kafka.id', '70%'),
-      new EntityTableColumn<KafkaConsumerGroup>('state', 'kafka.state', '10%'),
-      new EntityTableColumn<KafkaConsumerGroup>('members', 'kafka.members', '10%'),
-      new EntityTableColumn<KafkaConsumerGroup>('lag', 'kafka.lag', '10%', entity => entity.lag)
+      new EntityTableColumn<KafkaTopic>('name', 'kafka.name', '70%'),
+      new EntityTableColumn<KafkaTopic>('partitions', 'kafka.partitions', '10%',
+        undefined, () => ({color: 'rgba(0,0,0,0.54)'})),
+      new EntityTableColumn<KafkaTopic>('replicationFactor', 'kafka.replicas', '10%',
+        undefined, () => ({color: 'rgba(0,0,0,0.54)'})),
+      new EntityTableColumn<KafkaTopic>('size', 'kafka.size', '10%',
+          entity => formatBytes(entity.size),
+        () => ({color: 'rgba(0,0,0,0.54)'})
+      )
     );
   }
 
-  private fetchSessions(pageLink: TimePageLink): Observable<PageData<KafkaConsumerGroup>> {
-    return this.kafkaService.getKafkaConsumerGroups(pageLink);
+  private fetchKafkaTopics(pageLink: TimePageLink): Observable<PageData<KafkaTopic>> {
+    return this.kafkaService.getKafkaTopics(pageLink);
   }
 }
