+++ ui-ngx/src/app/modules/home/components/quick-links/quick-links.component.scss
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-@import "../../../../../scss/constants";
+@import 'constants';
 
 :host {
   .mat-button:disabled {
@@ -45,8 +45,8 @@
 
     .icon-container {
       display: flex;
-      width: 40px;
-      height: 40px;
+      width: 30px;
+      height: 30px;
       border-radius: 6px;
       align-items:center;
       justify-content:center;
