+++ ui-ngx/src/app/modules/home/pages/profile/profile.component.scss
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-@import "../../../../../scss/constants";
+@import 'constants';
 
 :host {
   mat-card.profile-card {
