+++ ui-ngx/src/app/shared/components/time/timeinterval.component.scss
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-@import '../../../../scss/constants';
+@import 'constants';
 
 :host {
   min-width: 355px;
